SSCS-9747 checkstyle
