+++ src/main/java/uk/gov/hmcts/reform/sscs/util/OtherPartyDataUtil.java
@@ -6,12 +6,10 @@ import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.isYes;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
-import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import uk.gov.hmcts.reform.sscs.ccd.domain.CcdValue;
 import uk.gov.hmcts.reform.sscs.ccd.domain.OtherParty;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.reform.sscs.ccd.domain.State;
 
 
 public class OtherPartyDataUtil {
@@ -42,19 +40,6 @@ public class OtherPartyDataUtil {
             if (otherParty.getRep() != null && isYes(otherParty.getRep().getHasRepresentative()) && otherParty.getRep().getId() == null) {
                 otherParty.getRep().setId(Integer.toString(++maxId));
             }
-        }
-    }
-
-    public static boolean hasDueDateSetAndOtherPartyWithoutHearingOption(SscsCaseData sscsCaseData) {
-        return StringUtils.isNotBlank(sscsCaseData.getDwpDueDate())
-                && !everyOtherPartyHasAtLeastOneHearingOption(sscsCaseData);
-    }
-
-    private static boolean everyOtherPartyHasAtLeastOneHearingOption(SscsCaseData sscsCaseData) {
-        if (sscsCaseData.getOtherParties() != null) {
-            return sscsCaseData.getOtherParties().stream().noneMatch(otherParty -> otherParty.getValue().getHearingOptions() == null);
-        } else {
-            return false;
         }
     }
