+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/confirmpanelcomposition/ConfirmPanelCompositionAboutToSubmitHandlerTest.java
@@ -8,8 +8,6 @@ import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.openMocks;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
 
-import java.time.LocalDate;
-import java.util.Arrays;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 import org.junit.Before;
