+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/updateotherparty/UpdateOtherPartySubmittedHandlerTest.java
@@ -1,14 +1,14 @@
-package uk.gov.hmcts.reform.sscs.ccd.presubmit.confirmpanelcomposition;
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.updateotherparty;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.BDDMockito.given;
+import static org.hamcrest.Matchers.nullValue;
+import static org.junit.Assert.*;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.openMocks;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.SUBMITTED;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
 
 import java.time.LocalDate;
 import java.util.Arrays;
@@ -22,23 +22,12 @@ import org.mockito.Mock;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.sscs.ccd.domain.*;
-import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
-import uk.gov.hmcts.reform.sscs.idam.IdamService;
-import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
 
 
 @RunWith(JUnitParamsRunner.class)
-public class ConfirmPanelCompositionSubmittedHandlerTest {
-
+public class UpdateOtherPartySubmittedHandlerTest {
     private static final String USER_AUTHORISATION = "Bearer token";
-    ConfirmPanelCompositionAboutToSubmitHandler handler;
-
-    private SscsCaseData sscsCaseData;
-
-    @Mock
-    private CcdService ccdService;
-    @Mock
-    private IdamService idamService;
+    private UpdateOtherPartyAboutToSubmitHandler handler;
 
     @Mock
     private Callback<SscsCaseData> callback;
@@ -46,13 +35,17 @@ public class ConfirmPanelCompositionSubmittedHandlerTest {
     @Mock
     private CaseDetails<SscsCaseData> caseDetails;
 
+    private SscsCaseData sscsCaseData;
+
+
     @Before
     public void setUp() {
         openMocks(this);
-        
+        handler = new UpdateOtherPartyAboutToSubmitHandler();
+
         sscsCaseData = SscsCaseData.builder().ccdCaseId("ccdId").build();
 
-        when(callback.getEvent()).thenReturn(EventType.CONFIRM_PANEL_COMPOSITION);
+        when(callback.getEvent()).thenReturn(EventType.UPDATE_OTHER_PARTY_DATA);
         when(callback.getCaseDetails()).thenReturn(caseDetails);
         when(caseDetails.getCaseData()).thenReturn(sscsCaseData);
     }
@@ -72,14 +65,11 @@ public class ConfirmPanelCompositionSubmittedHandlerTest {
         sscsCaseData.setDwpDueDate(LocalDate.now().toString());
         sscsCaseData.setOtherParties(Arrays.asList(buildOtherPartyWithHearing("2"), buildOtherPartyWithHearing("1")));
 
-        given(ccdService.updateCase(any(SscsCaseData.class), anyLong(), eq("readyToList"),
-                anyString(), anyString(), any(IdamTokens.class)))
-                .willReturn(SscsCaseDetails.builder().data(SscsCaseData.builder().state(State.READY_TO_LIST).build()).build());
-
         PreSubmitCallbackResponse<SscsCaseData> response =
-                handler.handle(SUBMITTED, callback, USER_AUTHORISATION);
+                handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
 
         assertThat(response.getData().getState(), is(State.READY_TO_LIST));
+        assertThat(response.getData().getDwpDueDate(), is(nullValue()));
     }
 
     @Test
@@ -127,16 +117,71 @@ public class ConfirmPanelCompositionSubmittedHandlerTest {
         assertThat(response.getData().getState(), is(State.NOT_LISTABLE));
     }
 
+    @Test
+    @Ignore
+    public void givenNoFqpmSetAndNoDueDateSetAndNotAllOtherPartyHearingOptionsSet_thenCaseStateIsNotListable() {
+
+        sscsCaseData.setOtherParties(Arrays.asList(buildOtherPartyWithHearing("2"), buildOtherParty("1", null)));
+
+        PreSubmitCallbackResponse<SscsCaseData> response =
+                handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
+
+        assertThat(response.getData().getState(), is(State.NOT_LISTABLE));
+    }
+
+    @Test
+    @Ignore
+    public void givenNoFqpmSetAndDueDateSetAndNotAllOtherPartyHearingOptionsSet_thenCaseStateIsNotListable() {
+
+        sscsCaseData.setOtherParties(Arrays.asList(buildOtherPartyWithHearing("2"), buildOtherParty("1", null)));
+        sscsCaseData.setDwpDueDate(LocalDate.now().toString());
+
+        PreSubmitCallbackResponse<SscsCaseData> response =
+                handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
+
+        assertThat(response.getData().getState(), is(State.NOT_LISTABLE));
+    }
+
+    @Test
+    @Ignore
+    public void givenNoFqpmSetAndNoDueDateSetAndAllOtherPartyHearingOptionsSet_thenCaseStateIsNotListable() {
+
+        sscsCaseData.setOtherParties(Arrays.asList(buildOtherPartyWithHearing("2"), buildOtherPartyWithHearing("1")));
+
+        PreSubmitCallbackResponse<SscsCaseData> response =
+                handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
+
+        assertThat(response.getData().getState(), is(State.NOT_LISTABLE));
+    }
+
     private CcdValue<OtherParty> buildOtherPartyWithHearing(String id) {
         return buildOtherParty(id, HearingOptions.builder().excludeDates(Arrays.asList(ExcludeDate.builder().build())).build());
+    }
+
+    private CcdValue<OtherParty> buildOtherParty(String id) {
+        return buildOtherParty(id, null);
     }
 
     private CcdValue<OtherParty> buildOtherParty(String id, HearingOptions hearingOptions) {
         return CcdValue.<OtherParty>builder()
                 .value(OtherParty.builder()
                         .id(id)
-                        .unacceptableCustomerBehaviour(YesNo.YES)
                         .hearingOptions(hearingOptions)
-                        .build()).build();
+                        .unacceptableCustomerBehaviour(YesNo.YES)
+                        .build())
+                .build();
     }
+
+    private CcdValue<OtherParty> buildOtherPartyWithAppointeeAndRep(String id, String appointeeId, String repId) {
+        return CcdValue.<OtherParty>builder()
+                .value(OtherParty.builder()
+                        .id(id)
+                        .isAppointee(YES.getValue())
+                        .appointee(Appointee.builder().id(appointeeId).build())
+                        .rep(Representative.builder().id(repId).hasRepresentative(YES.getValue()).build())
+                        .build())
+                .build();
+    }
+
+
 }
